Refactored .csv parser. Package name changed
