+++ pom.xml
@@ -9,13 +9,13 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.ss</groupId>
-	<artifactId>mail_sender</artifactId>
+	<artifactId>mailsender</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-<!--
+
 	<packaging>war</packaging>
-	<name>mail_sender</name>
+	<name>mailsender</name>
 	<description>Mail sender</description>
--->
+
 
 	<properties>
 		<java.version>11</java.version>
@@ -25,6 +25,7 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
@@ -36,6 +37,12 @@
 			<optional>true</optional>
 		</dependency>
 
+		<!-- https://mvnrepository.com/artifact/com.opencsv/opencsv -->
+		<dependency>
+			<groupId>com.opencsv</groupId>
+			<artifactId>opencsv</artifactId>
+			<version>5.3</version>
+		</dependency>
 		<!--
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
