--- src/main/java/com/ss/mail_sender/libs/CsvParser.java
@@ -1,33 +1,0 @@
-package com.ss.mail_sender.libs;
-
-import java.io.BufferedReader;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-
-public class CsvParser {
-
-    //Creating list of lines
-    List<String[]> data = new ArrayList<>();
-
-    //Setting delimiter
-    String delimiter = ";";
-
-    public List<String[]> csvParser(String fileName){
-
-        try(BufferedReader br = new BufferedReader(new FileReader(fileName))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] values = line.split(delimiter);
-                data.add(values);
-            }
-        } catch (FileNotFoundException e) {
-            System.out.println("File not found!");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return data;
-    }
-}
