--- src/main/java/com/ss/mail_sender/services/TestService.java
@@ -1,4 +1,0 @@
-package com.ss.mail_sender.services;
-
-public interface TestService {
-}
