+++ src/main/java/com/ss/mailsender/services/EmailSenderFileProcessor.java
@@ -1,11 +1,10 @@
-package com.ss.mail_sender.services;
+package com.ss.mailsender.services;
 
-import com.ss.mail_sender.model.Email;
-import com.ss.mail_sender.model.ProcessingResult;
-import com.ss.mail_sender.model.UploadingStatus;
+import com.ss.mailsender.model.Email;
+import com.ss.mailsender.model.ProcessingResult;
+import com.ss.mailsender.model.UploadingStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
