+++ src/main/java/com/ss/mailsender/services/TestService.java
@@ -1,0 +1,4 @@
+package com.ss.mailsender.services;
+
+public interface TestService {
+}
